+++ examples/wechat/src/example/observers/observers.js
@@ -1,0 +1,10 @@
+Page({
+  data: {
+    count: 9
+  },
+  updateCompProps() {
+    this.setData({
+      count: this.data.count + 1
+    })
+  }
+})