+++ packages/plugin-compiler-web/src/plugins/generateJSXEntryPlugin.ts
@@ -67,18 +67,19 @@ function generatePageJSX(
   showBack: boolean | string[] | ((pagePath: string) => boolean) = false,
   showHeader: boolean | string[] | ((pagePath: string) => boolean) = true
 ): string {
+  const page = slash(pagePath)
   return `import '${WEB_RUNTIMES.components}';
 import '${WEB_RUNTIMES.api}';
 import React from 'react';
-import EntryPage from '@/${pagePath}';
+import EntryPage from '@/${page}';
 
 class Page extends React.Component {
   render() {
     return (
       <tiga-page-host show-back="${shouldEnableFor(
         showBack,
-        pagePath
-      )}" show-header="${shouldEnableFor(showHeader, pagePath)}">
+        page
+      )}" show-header="${shouldEnableFor(showHeader, page)}">
         <EntryPage />
       </tiga-page-host>
     );
@@ -349,9 +350,7 @@ export class GenerateJSXEntryPlugin implements Plugin {
       // 处理分包里面的路径
       if (subPackageJson.root && subPackageJson.pages) {
         ;(subPackageJson.pages || []).map((page: string) => {
-          pages.push(
-            addLeadingSlash(slash(path.join(subPackageJson.root, page)))
-          )
+          pages.push(addLeadingSlash(slash(page)))
         })
       }
 
@@ -384,13 +383,18 @@ export class GenerateJSXEntryPlugin implements Plugin {
     // 生成各个页面对应的 jsx 文件
     pages.map((page) => {
       const pagePath = path.join(entriesRoot, `${page}.jsx`)
-      const pageContent = generatePageJSX(
-        entryHelpers.entries.get(page.slice(1)).entryName,
-        web.showBack,
-        web.showHeader
-      )
-      wrapper.fs.mem.mkdirpSync(path.dirname(pagePath))
-      wrapper.fs.mem.writeFileSync(pagePath, pageContent)
+      const entry = entryHelpers.entries.get(path.normalize(page.slice(1)))
+      if (entry?.entryName) {
+        const pageContent = generatePageJSX(
+          entry.entryName,
+          web.showBack,
+          web.showHeader
+        )
+        wrapper.fs.mem.mkdirpSync(path.dirname(pagePath))
+        wrapper.fs.mem.writeFileSync(pagePath, pageContent)
+      } else {
+        logger.warnOnce(`页面 ${page} 的 jsx 文件生成失败，原因：未找到文件`)
+      }
     })
 
     return { [MOR_GLOBAL_FILE(globalNameSuffix)]: mainEntryPath }