+++ packages/runtime-mini/src/alipay/componentToAlipay.ts
@@ -7,6 +7,7 @@ import {
   canIUse,
   injectComponentSelectorMethodsSupport,
   injectCreateIntersectionObserverSupport,
+  injectCreateSelectorQuerySupport,
   injectTwoWayBindingMethodsSupport,
   markUnsupportMethods
 } from './utilsToAlipay'
@@ -166,7 +167,7 @@ function injectEventSupport(options: Record<string, any>) {
       : {}
 
     // 支付宝组件传递函数时 必须以 on 开头并且 on 后的第一个字母必须大写（微信必须全小写）
-    if (typeof this.props[`on${name}`] === 'function') {
+    const callEventHandler = (eventName: string)=> {
       const currentTarget = eventObj.currentTarget || {}
       const target = eventObj.target || {}
       const e = {
@@ -178,7 +179,8 @@ function injectEventSupport(options: Record<string, any>) {
           },
           target: {
             ...target,
-            dataset: { ...(target.dataset || {}), ...dataset }
+            dataset: { ...(target.dataset || {}), ...dataset },
+            id: this.props.id
           }
         },
         type: name
@@ -194,7 +196,12 @@ function injectEventSupport(options: Record<string, any>) {
         e.detail = params
       }
 
-      this.props[`on${name}`](e, opts)
+      this.props[eventName](e, opts)
+    }
+    if (typeof this.props[`on${name}`] === 'function') {
+      callEventHandler(`on${name}`)
+    } else if (typeof this.props[`catch${name}`] === 'function') {
+      callEventHandler(`catch${name}`)
     }
   }
 
@@ -348,10 +355,10 @@ function injectPropertiesAndObserversSupport(options: Record<string, any>) {
     // 1. 当基础库版本支持 lifetimes 时，由于生命周期执行委托给了原生，需跳过首次执行，若不跳过则会导致，
     //    data 同步 nextProps 后，传入的值前后对比未发现变更，而使在第一次初始化不触发 observer 的监听
     // 2. 当基础库版本不支持 lifetimes 时，使用 mor 的自实现，正常执行以下流程
-    if (!this[MOR_FIRST_DERIVE_DATA_FROM_PROPS] && isObserversSupported) {
-      this[MOR_FIRST_DERIVE_DATA_FROM_PROPS] = true
-      return
-    }
+    // 基于上述逻辑，初始化时，需记录状态，跳过 properties/data 赋值
+    const firstDeriveDataFromProps = !this[MOR_FIRST_DERIVE_DATA_FROM_PROPS]
+    const firstDeriveWithObserversSupported =
+      firstDeriveDataFromProps && isObserversSupported
 
     // data变化触发双向绑定
     this.props.onMorChildTWBProxy?.(this.data, this.props)
@@ -375,15 +382,24 @@ function injectPropertiesAndObserversSupport(options: Record<string, any>) {
 
       const originalValue = this.properties[prop]
 
-      // 更新 properties 和 data
-      // 微信小程序中的 properties 和 data 是一致的
-      // 都包含 包括内部数据和属性值
-      this.properties[prop] = nextProps[prop]
-      this.data[prop] = nextProps[prop]
+      // 基于上述 1、2 两点逻辑，支持 observers 的版本初始化时不赋值。
+      // 但是该做法会导致 property.observer 首次回调中，this.data
+      // 和 this.properties 值不同步，可建议用户首次避免通过 data
+      // 获取 properties 中的属性值
+      if (!firstDeriveWithObserversSupported) {
+        // 更新 properties 和 data
+        // 微信小程序中的 properties 和 data 是一致的
+        // 都包含 包括内部数据和属性值
+        this.properties[prop] = nextProps[prop]
+        this.data[prop] = nextProps[prop]
+      }
 
-      // 执行属性监听器，仅执行发生了变化的属性
+      // 微信端组件初始化、属性改变时，会触发属性监听器 property.observer。
+      // 而支付宝 deriveDataFromProps 方法，初始化时 this.props 和 nextProps
+      // 中属性值相同，不满足 isPropChanged = true
+      // 需通过 firstDeriveDataFromProps 强制初始化触发，与微信逻辑同步
       if (
-        isPropChanged &&
+        (isPropChanged || firstDeriveDataFromProps) &&
         propertiesWithObserver[prop] &&
         !(pureDataPattern && pureDataPattern.test(prop))
       ) {
@@ -401,6 +417,11 @@ function injectPropertiesAndObserversSupport(options: Record<string, any>) {
         }
       }
     }
+
+    // 初始化时，记录首次执行状态
+    this[MOR_FIRST_DERIVE_DATA_FROM_PROPS] = true
+    if (firstDeriveWithObserversSupported) return
+
     // 触发一次更新
     if (hasProps) {
       this.setData(updateProps)
@@ -484,6 +505,7 @@ function hookComponentLifeCycle(options: Record<string, any>) {
   options.onInit = compose([
     hackSetData,
     injectCreateIntersectionObserverSupport(),
+    injectCreateSelectorQuerySupport(),
     initPropertiesAndData,
     callOriginalFn('onInit')
   ])